# Used ArrayPool and `yield return` instead of `new Span`/`new List`