This change significantly improves memory usage of `Query` calls when a lot of colliding ranges are found. 
The benchmarks from this repository don't show any improvements for some reason, but the following one shows quite a huge improvement:

```cs

[MemoryDiagnoser]
public class IntervalTreeBenchmarks
{
    private readonly QuickIntervalTree<uint, uint> _quickIntervalTree = new();
    private uint[] _points = new uint[10_000];

    [GlobalSetup]
    public void Setup()
    {
        for (int i = 0; i < _points.Length; ++i)
        {
            uint low = (uint)Random.Shared.Next(10_000);
            uint high = (uint)Random.Shared.Next((int)low + 1, 20_000);

            _quickIntervalTree.Add(low, high, (uint)Random.Shared.Next(20_000));
            _points[i] = (uint)Random.Shared.Next();
        }
    }

    [Benchmark(Baseline = true)]
    public void Old()
    {
        for (uint i = 0; i < _points.Length; ++i)
        {
            int count = 0;
            foreach (uint u in _quickIntervalTree.Query(i))
            {
                ++count;
            }
        }
    }

    [Benchmark]
    public void New()
    {
        for (uint i = 0; i < _points.Length; ++i)
        {
            int count = 0;
            foreach (uint u in _quickIntervalTree.QueryNew(i))
            {
                ++count;
            }
        }
    }
}
```
Results:
| Method | Mean     | Error   | StdDev  | Ratio | RatioSD | Gen0       | Gen1      | Allocated    | Alloc Ratio |
|------- |---------:|--------:|--------:|------:|--------:|-----------:|----------:|-------------:|------------:|
| Old    | 142.2 ms | 2.64 ms | 2.71 ms |  1.00 |    0.03 | 36750.0000 | 3750.0000 | 450985.72 KB |       1.000 |
| New    | 131.5 ms | 1.51 ms | 1.18 ms |  0.92 |    0.02 |          - |         - |    781.41 KB |       0.002 |


Maybe there is something wrong with my benchmark code and the code I'm proposing in the PR isn't making any difference, but I couldn't find why that would be the case.