@jamarino 2025-03-24
Hello @Egorikhin!

Thank you for your PR, I have been reviewing and testing your changes over the last few days.

I have a new feature coming soon, which also addresses allocations, but in a different way. So I have reviewed your changes with that in mind.

**Yield-return**
Using the yield-return pattern has some advantages, especially in cases where a query would return many results, but don't need all of them. This, however, does seem like a niche use case, especially since the return order is not guaranteed.

One drawback of yield-return is that it requires allocating a state machine - even if there are no results. Furthermore, it removes the ability to use stack-allocated buffers, which forces us to either 1) use a recursive search instead of a stack-based DFS, or 2) find an alternative buffer for the DFS stack.

**Recursive search**
One problem I ran into when converting the search algorithm to be recursive, is that a new state machine is allocated for each recursive call. So for each query performed, a large number of state machines may be allocated, just to avoid allocating a single output buffer. Maybe I have missed a trick here..

**Alternative buffer**
Simply allocating a new array per query, negates the idea of using yield-return to reduce allocations. Maybe in some cases the stack buffer is smaller than your output buffer, and so your memory writes are reduced, but this is highly speculative.

Using an ArrayPool to recycle the arrays is an interesting idea, but it has some caveats.

Firstly, ArrayPools do have some overhead, so they are most commonly recommended for re-using medium to larger buffers. (in my benchmarking, it was faster to just allocate new arrays `new int[]`)

Secondly, as with any shared resource pools, concurrency is a consideration. I don't know for sure what impact this may have in highly concurrent scenarios.

Thirdly, and most importantly, it seems that the combination of yield-return and ArrayPool may have introduced a memory leak. You call `ArrayPool.Return()` at the very end of the query method, but this code is not always reached.
If, for instance, I call `tree.Query(x).FirstOrDefault()`, there will be cases where the enumerator is not yet finished returning all results, before being discarded. In this case, the rented buffer is never returned, resulting in a memory-leak.
Please correct me if I am wrong about this analysis.

**Conclusion**
While the yield-return pattern does have some benefits, I think the drawbacks in this case are too severe. I am quite happy with performance and low-allocation nature of the current stack-based DFS algorithm.

As mentioned, I am releasing a new feature soon which addresses the allocation of the output buffer. It's a new query method, called `Find()`, which implements a fold- or aggregation-pattern.
Essentially, you give the `Find()` method a function to run for each matching interval found, and that function could simply add the interval to an output buffer, or calculate some metric on the fly (sum, max, avg) without ever needing an output buffer.
In my testing this is showing improved performance and in some cases completely eliminates allocations, which I am quite excited about.

Once again thank you for submitting this PR, I greatly appreciate the effort you have put in and the opportunity to reflect on some alternative approaches.

Kind regards,
Jonathan