@Egorikhin 2025-03-25
Hello @jamarino, thanks for such a detailed response!
Your concern about memory leak is correct, haven't thought about that, though easily fixable with try/finally with array returned inside the finally block.

>, however, does seem like a niche use case, especially since the return order is not guaranteed.

I'm not quite getting the issue with ordering: yield return will just return the item in the same order it was returned in an old approach, no change there. In my use case I wanted to find the first period in the interval with some properties so I used something like `tree.Query(123).FirstOrDefault(x => x.Property == true)` so I would have an improved perf by not having to buffer the whole result .

Regarding allocations, recursion and small buffers: 
My initial concern was about using a single tree to make many queries, returning many overlapping periods. In such scenario allocating a list and appending items to it results in several memory copies to enlarge  the underlying array leading to the results presented in the benchmark. 
Using yield return here is a tradeoff, that's right, we're trading the constant upfront cost of allocating the state machine to potential cost of allocating expandable list of unknown number of items (zero in the best case). 
Wouldn't you mind sharing the DFS approach you're developing, maybe there is a possibility to remove recursion and make it iterative using yield return approach to have the benefits of lazy evaluations? It's also possible that using plain old array allocation (not span as it isn't working with Spans) with the small constant number of elements would be enough and  ArrayPool usage could be removed entirely.

Thanks again for your response, it was a pleasure to read :)